Create framework for leveraging user profiles
Add support for storing 'profiles', their corresponding
'attrs', and for linking profiles to users through the
'profiles_users' table.

Profiles may also be linked to users temporarily through
the 'SessionProfileManager', which adds certain profiles
to a user scope session. All profiles to which a user is
linked are exposed through the 'ProfileAttributes' interface,
which merges the attributes of user and session profiles
into a single hash containing all profile attributes.

Profile attributes support use cases such as the configuring
of default overlays, which is exemplified in the 'Overlays'
class.