Fix image export in editor
This commit changes the behaviour of the app so it will embed
auth_tokens in the admin visualization show erb when the logged in user
has read access to the visualizations. It also changes the UI library so
it will know to pass these tokens to the StaticImage instantiation that
wraps the Maps API static image endpoint. This request won't fail now
since the appropiate tokens are available (required for private
visualizations).