Make export image download interactive
In Chromium versions 64+, the behavior of the popup
blocker changes such that the export image download
is now blocked by default if popups from the site's
domain are not whitelisted. This occurs when the
javascript which launches the popup is not a direct
result of a user action. Previously, the download
link was clicked automatically in JavaScript when the
download was ready. This change exposes a link on
which the user can click to download the image, thus
preventing Chrome from blocking the popup.