Open exported image in current tab by default
The previous behavior of the export image dialog
was to open the exported image for download in a
new tab by specifying target="_blank" on the down-
load link.  This worked fine before Chrome 65, at
which point the behavior changed such that the
popup was blocked by default, and triggered a new
code path in the chromium wrapper for handling
links which was subject to crashes.  Removing
target="_blank" from the download link forces the
download to be processed in the current tab, and
bypasses the problematic code path in our chromium
wrapper. This change may be backed by creating the
unrestricted feature flag 'export_image_in_new_tab'.